--- a/Misc/NEWS.d/next/Library/2025-01-29-10-53-32.gh-issue-118761.i8wjpV.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-29-10-53-32.gh-issue-118761.i8wjpV.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`subprocess` by lazy importing ``locale`` and
+``signal``. Patch by Taneli Hukkinen.