This makes `import subprocess` lazy import `locale` and `signal`. As a result, also `re` (imported by `locale`) and `enum` (imported by `signal`) imports are skipped.

## Benchmark import time

### `main` branch

```console
$ hyperfine --warmup 11 "./python -c 'import subprocess'" "./python -c 'pass'"
Benchmark 1: ./python -c 'import subprocess'
  Time (mean ± σ):      24.6 ms ±   0.9 ms    [User: 20.8 ms, System: 4.0 ms]
  Range (min … max):    23.5 ms …  28.0 ms    108 runs
 
Benchmark 2: ./python -c 'pass'
  Time (mean ± σ):      10.8 ms ±   0.6 ms    [User: 8.9 ms, System: 2.3 ms]
  Range (min … max):    10.1 ms …  14.2 ms    190 runs
 
Summary
  './python -c 'pass'' ran
    2.28 ± 0.16 times faster than './python -c 'import subprocess''
```

### PR branch

```console
$ hyperfine --warmup 11 "./python -c 'import subprocess'" "./python -c 'pass'"
Benchmark 1: ./python -c 'import subprocess'
  Time (mean ± σ):      18.7 ms ±   0.8 ms    [User: 16.1 ms, System: 3.2 ms]
  Range (min … max):    17.6 ms …  21.7 ms    139 runs
 
Benchmark 2: ./python -c 'pass'
  Time (mean ± σ):      10.4 ms ±   0.5 ms    [User: 9.4 ms, System: 1.8 ms]
  Range (min … max):     9.7 ms …  12.9 ms    226 runs
 
Summary
  './python -c 'pass'' ran
    1.80 ± 0.11 times faster than './python -c 'import subprocess''
```

## Benchmark `subprocess.run` call 
In fact, this not only has an effect on import time, but simple `subprocess.run` calls, too:

### `main` branch

```console
$ hyperfine --warmup 11 "./python -c 'import subprocess; subprocess.run([\"sleep\", \"0\"])'" "./python -c 'pass'"
Benchmark 1: ./python -c 'import subprocess; subprocess.run(["sleep", "0"])'
  Time (mean ± σ):      26.7 ms ±   1.1 ms    [User: 23.4 ms, System: 3.4 ms]
  Range (min … max):    25.5 ms …  31.7 ms    103 runs
 
Benchmark 2: ./python -c 'pass'
  Time (mean ± σ):      11.7 ms ±   0.8 ms    [User: 9.5 ms, System: 2.4 ms]
  Range (min … max):    10.9 ms …  15.4 ms    235 runs
 
Summary
  './python -c 'pass'' ran
    2.28 ± 0.18 times faster than './python -c 'import subprocess; subprocess.run(["sleep", "0"])''
```

### PR branch

```console
$ hyperfine --warmup 11 "./python -c 'import subprocess; subprocess.run([\"sleep\", \"0\"])'" "./python -c 'pass'"
Benchmark 1: ./python -c 'import subprocess; subprocess.run(["sleep", "0"])'
  Time (mean ± σ):      22.0 ms ±   1.3 ms    [User: 18.1 ms, System: 3.9 ms]
  Range (min … max):    20.3 ms …  27.2 ms    119 runs
 
Benchmark 2: ./python -c 'pass'
  Time (mean ± σ):      12.0 ms ±   0.4 ms    [User: 9.8 ms, System: 2.3 ms]
  Range (min … max):    11.5 ms …  13.6 ms    224 runs
 
Summary
  './python -c 'pass'' ran
    1.84 ± 0.12 times faster than './python -c 'import subprocess; subprocess.run(["sleep", "0"])''
```


<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->